+++ packages/headless/src/case-assist.index.ts
@@ -30,3 +30,17 @@ export * from './features/case-input/case-input-actions-loader';
 export * from './features/case-field/case-field-actions-loader';
 export * from './features/document-suggestion/document-suggestion-actions-loader';
 export * from './features/case-assist/case-assist-analytics-actions-loader';
+
+// Controllers
+export {
+  Controller,
+  buildController,
+} from './controllers/controller/headless-controller';
+
+export {
+  CaseInputState,
+  CaseInput,
+  CaseInputOptions,
+  CaseInputProps,
+  buildCaseInput,
+} from './controllers/case-input/headless-case-input';