+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.38.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.37.2...@coveo/headless@1.38.0) (2021-12-01)
+
+
+### Features
+
+* **headless:** added case input controller ([#1542](https://github.com/coveo/ui-kit/issues/1542)) ([9c7cb42](https://github.com/coveo/ui-kit/commit/9c7cb420c68ab890aac19bd1363b1c154e722767))
+
+
+
+
+
 ## [1.37.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.37.1...@coveo/headless@1.37.2) (2021-11-30)
 
 **Note:** Version bump only for package @coveo/headless