+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.1.2...@coveo/quantic@1.1.3) (2021-12-01)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [1.1.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.1.1...@coveo/quantic@1.1.2) (2021-11-30)
 
 **Note:** Version bump only for package @coveo/quantic